feat: useToken
