+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "lets-onion",
   "version": "1.0.0",
+  "private": true,
   "main": "expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
@@ -31,6 +32,5 @@
     "@types/react": "~18.2.45",
     "@types/styled-components": "^5.1.34",
     "typescript": "^5.1.3"
-  },
-  "private": true
+  }
 }
